lemon refactor iterate
